refactor package within Schema to v0_1
