+++ generator/src/main/java/com/scottlogic/deg/generator/decisiontree/serialisation/DecisionTreeMapper.java
@@ -5,7 +5,7 @@ import com.scottlogic.deg.generator.ProfileFields;
 import com.scottlogic.deg.generator.inputs.RuleInformation;
 import com.scottlogic.deg.generator.constraints.atomic.*;
 import com.scottlogic.deg.generator.decisiontree.*;
-import com.scottlogic.deg.schemas.v3.RuleDTO;
+import com.scottlogic.deg.schemas.v0_1.RuleDTO;
 
 import java.util.*;
 import java.util.stream.Collectors;
