+++ generator/src/main/java/com/scottlogic/deg/generator/inputs/ConstraintReader.java
@@ -2,7 +2,7 @@ package com.scottlogic.deg.generator.inputs;
 
 import com.scottlogic.deg.generator.ProfileFields;
 import com.scottlogic.deg.generator.constraints.Constraint;
-import com.scottlogic.deg.schemas.v3.ConstraintDTO;
+import com.scottlogic.deg.schemas.v0_1.ConstraintDTO;
 
 import java.util.Set;
