+++ generator/src/main/java/com/scottlogic/deg/generator/inputs/JsonProfileReader.java
@@ -7,7 +7,7 @@ import com.scottlogic.deg.generator.ProfileFields;
 import com.scottlogic.deg.generator.Rule;
 import com.scottlogic.deg.generator.inputs.validation.ProfileValidator;
 import com.scottlogic.deg.schemas.common.ProfileDeserialiser;
-import com.scottlogic.deg.schemas.v3.ProfileDTO;
+import com.scottlogic.deg.schemas.v0_1.ProfileDTO;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
