+++ generator/src/test/java/com/scottlogic/deg/generator/builders/RuleBuilder.java
@@ -1,15 +1,8 @@
 package com.scottlogic.deg.generator.builders;
 
-import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.Rule;
-import com.scottlogic.deg.generator.constraints.Constraint;
-import com.scottlogic.deg.generator.constraints.atomic.IsGreaterThanConstantConstraint;
-import com.scottlogic.deg.generator.constraints.atomic.IsLessThanConstantConstraint;
 import com.scottlogic.deg.generator.inputs.RuleInformation;
-import com.scottlogic.deg.schemas.v3.RuleDTO;
-
-import java.util.ArrayList;
-import java.util.Collection;
+import com.scottlogic.deg.schemas.v0_1.RuleDTO;
 
 /**
  * Defines a builder for a Rule object
