+++ generator/src/test/java/com/scottlogic/deg/generator/constraints/grammatical/AndConstraintTests.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.constraints.grammatical;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.constraints.atomic.IsNullConstraint;
 import com.scottlogic.deg.generator.inputs.RuleInformation;
-import com.scottlogic.deg.schemas.v3.RuleDTO;
+import com.scottlogic.deg.schemas.v0_1.RuleDTO;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
