+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/steps/TypeRegistryConfiguration.java
@@ -5,7 +5,7 @@ import com.scottlogic.deg.generator.cucumber.utils.CucumberGenerationMode;
 import com.scottlogic.deg.generator.cucumber.utils.GeneratorTestUtilities;
 import com.scottlogic.deg.generator.generation.GenerationConfig;
 import com.scottlogic.deg.generator.inputs.InvalidProfileException;
-import com.scottlogic.deg.schemas.v3.AtomicConstraintType;
+import com.scottlogic.deg.schemas.v0_1.AtomicConstraintType;
 import cucumber.api.TypeRegistry;
 import cucumber.api.TypeRegistryConfigurer;
 import io.cucumber.cucumberexpressions.ParameterType;
