+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/utils/CucumberGenerationConfigSource.java
@@ -3,7 +3,7 @@ package com.scottlogic.deg.generator.cucumber.utils;
 import com.google.inject.Inject;
 import com.scottlogic.deg.generator.generation.GenerationConfig;
 import com.scottlogic.deg.generator.generation.GenerationConfigSource;
-import com.scottlogic.deg.schemas.v3.AtomicConstraintType;
+import com.scottlogic.deg.schemas.v0_1.AtomicConstraintType;
 
 import java.io.File;
 import java.nio.file.Path;
