+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/utils/CucumberProfileReader.java
@@ -9,7 +9,7 @@ import com.scottlogic.deg.generator.inputs.InvalidProfileException;
 import com.scottlogic.deg.generator.inputs.MainConstraintReader;
 import com.scottlogic.deg.generator.inputs.ProfileReader;
 import com.scottlogic.deg.generator.inputs.RuleInformation;
-import com.scottlogic.deg.schemas.v3.RuleDTO;
+import com.scottlogic.deg.schemas.v0_1.RuleDTO;
 
 import java.nio.file.Path;
 import java.util.Collections;
