+++ generator/src/test/java/com/scottlogic/deg/generator/cucumber/utils/CucumberTestState.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.generation.GenerationConfig;
-import com.scottlogic.deg.schemas.v3.AtomicConstraintType;
-import com.scottlogic.deg.schemas.v3.ConstraintDTO;
+import com.scottlogic.deg.schemas.v0_1.AtomicConstraintType;
+import com.scottlogic.deg.schemas.v0_1.ConstraintDTO;
 
 import java.io.IOException;
 import java.util.*;
