+++ generator/src/test/java/com/scottlogic/deg/generator/generation/TestGenerationConfigSource.java
@@ -1,6 +1,6 @@
 package com.scottlogic.deg.generator.generation;
 
-import com.scottlogic.deg.schemas.v3.AtomicConstraintType;
+import com.scottlogic.deg.schemas.v0_1.AtomicConstraintType;
 
 import java.io.File;
 import java.nio.file.Path;
