+++ generator/src/test/java/com/scottlogic/deg/generator/inputs/AtomicConstraintReaderLookupTests.java
@@ -4,9 +4,9 @@ import com.scottlogic.deg.generator.Field;
 import com.scottlogic.deg.generator.ProfileFields;
 import com.scottlogic.deg.generator.constraints.Constraint;
 import com.scottlogic.deg.generator.constraints.atomic.*;
-import com.scottlogic.deg.schemas.v3.AtomicConstraintType;
-import com.scottlogic.deg.schemas.v3.ConstraintDTO;
-import com.scottlogic.deg.schemas.v3.RuleDTO;
+import com.scottlogic.deg.schemas.v0_1.AtomicConstraintType;
+import com.scottlogic.deg.schemas.v0_1.ConstraintDTO;
+import com.scottlogic.deg.schemas.v0_1.RuleDTO;
 import org.junit.Assert;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
