+++ schemas/src/main/java/com/scottlogic/deg/schemas/v0_1/UndefinedValue.java
@@ -1,0 +1,4 @@
+package com.scottlogic.deg.schemas.v0_1;
+
+public class UndefinedValue {
+}
