--- schemas/src/main/java/com/scottlogic/deg/schemas/v3/UndefinedValue.java
@@ -1,4 +1,0 @@
-package com.scottlogic.deg.schemas.v3;
-
-public class UndefinedValue {
-}
